server/app/requirements.txt
fastapi~=0.62.0
uvicorn~=0.13.1
pydantic~=1.7.3
sqlalchemy~=1.3.22
psycopg2-binary~=2.8.6
aiopg~=1.1.0